requirements.txt
watchdog==2.1.9
Pillow==9.2.0